feat(editor): pass all test-keyboardless tests except test_cand_list

+++ capi/chewing-internal/src/io.rs
@@ -151,7 +151,8 @@ pub extern "C" fn chewing_new2(
         keyboard,
         editor,
         kbcompat_iter: None,
-        cand_iter: None,
+        candidates: None,
+        candidate_cursor: 0,
         interval_iter: None,
     });
     Box::into_raw(context)
@@ -703,23 +704,35 @@ pub extern "C" fn chewing_userphrase_lookup(
 #[tracing::instrument(skip(ctx), ret)]
 #[no_mangle]
 pub extern "C" fn chewing_cand_list_first(ctx: *mut ChewingContext) -> c_int {
-    let ctx = match unsafe { ctx.as_ref() } {
+    let ctx = match unsafe { ctx.as_mut() } {
         Some(ctx) => ctx,
         None => return -1,
     };
 
-    todo!()
+    match ctx.candidates.as_ref() {
+        Some(_) => {
+            ctx.candidate_cursor = 0;
+            0
+        }
+        None => -1,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
 #[no_mangle]
 pub extern "C" fn chewing_cand_list_last(ctx: *mut ChewingContext) -> c_int {
-    let ctx = match unsafe { ctx.as_ref() } {
+    let ctx = match unsafe { ctx.as_mut() } {
         Some(ctx) => ctx,
         None => return -1,
     };
 
-    todo!()
+    match ctx.candidates.as_ref() {
+        Some(candidates) => {
+            ctx.candidate_cursor = candidates.len() - 1;
+            0
+        }
+        None => -1,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
@@ -730,7 +743,13 @@ pub extern "C" fn chewing_cand_list_has_next(ctx: *mut ChewingContext) -> c_int 
         None => return 0,
     };
 
-    todo!()
+    match ctx.candidates.as_ref() {
+        Some(candidates) => match candidates.get(ctx.candidate_cursor + 1) {
+            Some(_) => 1,
+            None => 0,
+        },
+        None => 0,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
@@ -741,36 +760,67 @@ pub extern "C" fn chewing_cand_list_has_prev(ctx: *mut ChewingContext) -> c_int 
         None => return 0,
     };
 
-    todo!()
+    match ctx.candidates.as_ref() {
+        Some(candidates) => match candidates.get(ctx.candidate_cursor - 1) {
+            Some(_) => 1,
+            None => 0,
+        },
+        None => 0,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
 #[no_mangle]
 pub extern "C" fn chewing_cand_list_next(ctx: *mut ChewingContext) -> c_int {
-    // FIXME selecting next mode
-    chewing_handle_Down(ctx)
+    let ctx = match unsafe { ctx.as_mut() } {
+        Some(ctx) => ctx,
+        None => return -1,
+    };
+    match ctx.candidates.as_ref() {
+        Some(candidates) => match candidates.get(ctx.candidate_cursor + 1) {
+            Some(_) => {
+                ctx.candidate_cursor += 1;
+                0
+            }
+            None => -1,
+        },
+        None => -1,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
 #[no_mangle]
 pub extern "C" fn chewing_cand_list_prev(ctx: *mut ChewingContext) -> c_int {
-    let ctx = match unsafe { ctx.as_ref() } {
+    let ctx = match unsafe { ctx.as_mut() } {
         Some(ctx) => ctx,
         None => return -1,
     };
-
-    todo!()
+    match ctx.candidates.as_ref() {
+        Some(candidates) => match candidates.get(ctx.candidate_cursor - 1) {
+            Some(_) => {
+                ctx.candidate_cursor -= 1;
+                0
+            }
+            None => -1,
+        },
+        None => -1,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
 #[no_mangle]
 pub extern "C" fn chewing_commit_preedit_buf(ctx: *mut ChewingContext) -> c_int {
-    let ctx = match unsafe { ctx.as_ref() } {
+    let ctx = match unsafe { ctx.as_mut() } {
         Some(ctx) => ctx,
         None => return -1,
     };
 
-    todo!()
+    // FIXME
+    if !ctx.editor.is_entering() || ctx.editor.display().is_empty() {
+        -1
+    } else {
+        chewing_handle_Enter(ctx)
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
@@ -781,8 +831,12 @@ pub extern "C" fn chewing_clean_preedit_buf(ctx: *mut ChewingContext) -> c_int {
         None => return -1,
     };
 
-    ctx.editor.clear();
-    0
+    if !ctx.editor.is_entering() {
+        -1
+    } else {
+        ctx.editor.clear();
+        0
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
@@ -851,7 +905,7 @@ pub extern "C" fn chewing_handle_Enter(ctx: *mut ChewingContext) -> c_int {
 
     ctx.editor
         .process_keyevent(ctx.keyboard.map(KeyCode::Enter));
-    1
+    0
 }
 
 #[tracing::instrument(skip(ctx), ret)]
@@ -1295,8 +1349,8 @@ pub extern "C" fn chewing_cand_Enumerate(ctx: *mut ChewingContext) {
     match ctx.editor.paginated_candidates() {
         Ok(candidates) => {
             debug!("candidates: {candidates:?}");
-            let phrases = Box::new(candidates.into_iter()) as Box<dyn Iterator<Item = String>>;
-            ctx.cand_iter = Some(phrases.peekable());
+            ctx.candidates = Some(candidates);
+            ctx.candidate_cursor = 0;
         }
         Err(_) => (),
     }
@@ -1310,10 +1364,13 @@ pub extern "C" fn chewing_cand_hasNext(ctx: *mut ChewingContext) -> c_int {
         None => return -1,
     };
 
-    ctx.cand_iter
-        .as_mut()
-        .and_then(|it| it.peek())
-        .map_or(0, |_| 1)
+    match ctx.candidates.as_ref() {
+        Some(candidates) => match candidates.get(ctx.candidate_cursor) {
+            Some(_) => 1,
+            None => 0,
+        },
+        None => 0,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
@@ -1324,12 +1381,17 @@ pub extern "C" fn chewing_cand_String(ctx: *mut ChewingContext) -> *mut c_char {
         None => return owned_into_raw(Owned::CString, CString::default().into_raw()),
     };
 
-    match ctx.cand_iter.as_mut().and_then(|it| it.next()) {
+    match ctx
+        .candidates
+        .as_ref()
+        .and_then(|candidates| candidates.get(ctx.candidate_cursor))
+    {
         Some(phrase) => {
-            let cstr = match CString::new(String::from(phrase)) {
+            let cstr = match CString::new(phrase.clone()) {
                 Ok(cstr) => cstr,
-                Err(_) => return null_mut(),
+                Err(_) => return owned_into_raw(Owned::CString, CString::default().into_raw()),
             };
+            ctx.candidate_cursor += 1;
             owned_into_raw(Owned::CString, cstr.into_raw())
         }
         None => owned_into_raw(Owned::CString, CString::default().into_raw()),
@@ -1344,8 +1406,15 @@ pub extern "C" fn chewing_cand_String_static(ctx: *mut ChewingContext) -> *const
         None => return unsafe { global_empty_cstr() },
     };
 
-    match ctx.cand_iter.as_mut().unwrap().next() {
-        Some(phrase) => unsafe { global_cstr(&String::from(phrase)) },
+    match ctx
+        .candidates
+        .as_ref()
+        .and_then(|candidates| candidates.get(ctx.candidate_cursor))
+    {
+        Some(phrase) => {
+            ctx.candidate_cursor += 1;
+            unsafe { global_cstr(phrase) }
+        }
         None => unsafe { global_empty_cstr() },
     }
 }
@@ -1386,19 +1455,29 @@ pub extern "C" fn chewing_cand_string_by_index_static(
 #[tracing::instrument(skip(ctx), ret)]
 #[no_mangle]
 pub extern "C" fn chewing_cand_choose_by_index(ctx: *mut ChewingContext, index: c_int) -> c_int {
-    let ctx = match unsafe { ctx.as_ref() } {
+    let ctx = match unsafe { ctx.as_mut() } {
         Some(ctx) => ctx,
         None => return -1,
     };
 
-    todo!()
+    match ctx.editor.select(index as usize) {
+        Ok(_) => 0,
+        Err(_) => -1,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
 #[no_mangle]
 pub extern "C" fn chewing_cand_open(ctx: *mut ChewingContext) -> c_int {
-    // FIXME enter selecting mode
-    chewing_handle_Down(ctx)
+    let ctx = match unsafe { ctx.as_mut() } {
+        Some(ctx) => ctx,
+        None => return -1,
+    };
+
+    match ctx.editor.start_selecting() {
+        Ok(_) => 0,
+        Err(_) => -1,
+    }
 }
 
 #[tracing::instrument(skip(ctx), ret)]
